Add files via upload
In this program, you are to store an array of student records. Each record will be stored
in a C++ object. Thus, you will define a C++ class called StudentRecord that will contain the
appropriate data. The data in this course will assume each student in the course is a freshman in
his/her second semester. 
Once you have read this data into the array, you will go into a loop. This loop will
terminate when the input character is ‘d’. During this loop, you will query the user for the
following command:

1. print list of all student names
2. print list of all student names + current GPA (from last semester) of each student
3. select students by last name, print the information about all students with the
same last name. For example, type “smith” and print out the information about
all students named “smith”
4. select a student by name, print the information about this one student
a. You will have to select a student by both first name and last name for
this.

5. select a student, calculate the student’s new GPA based on his/her old GPA plus
the grades from this semester
6. print list of all student names + student’s new GPA based on his/her old GPA
plus the grades from this semester